requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
werkzeug==3.1.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
python-decouple==3.8
flask-cors==5.0.1
flasgger==0.9.7.1
requests
